+++ index.js
@@ -1,6 +1,6 @@
 const SPECIAL = ['[', '^', '$', '.', '|', '?', '*', '+', '(', ')'];
 
-class Split {
+const Split = {
   /**
    * The base method of the split-it library.
    * Splits a given string on all "classic" delimiters:
@@ -8,9 +8,9 @@ class Split {
    *
    * @param {string} str the string to split
    */
-  it = (str) => {
+  it : (str) => {
     return str.split(/,|:|\t|\r\n|\r|\n/g);
-  };
+  },
 
   /**
    * Accepts an optional delimiter argument.
@@ -18,7 +18,7 @@ class Split {
    * @param {string} str the string to split
    * @param {string} del the chosen delimiter on which to split the given string
    */
-  onDelimiter = (str, del = null) => {
+  onDelimiter : (str, del = null) => {
     if (del) {
       let escape = SPECIAL.includes(del) ? '\\' : '';
       let re = new RegExp(escape + del, 'g');
@@ -26,7 +26,7 @@ class Split {
     } else {
       return str.split(/\s/g);
     }
-  };
+  },
 
   /**
    *
@@ -34,7 +34,7 @@ class Split {
    * @param options an object containing boolean values for "headings" and "splitOnColumns"
    * @return {string[]} the passed in csv data will be split into an array of arrays, with each sub-array representing a single row or column of data. 
    */
-  csv = (data, { headings = true, splitOnColumns = false } = {}) => {
+  csv : (data, { headings = true, splitOnColumns = false } = {}) => {
     let rows = data.split(/\r\n|\n|\r/g).map((row) => row.split(/,\s+/g));
     let returnData = [];
 
@@ -58,7 +58,7 @@ class Split {
     }
 
     return returnData;
-  };
+  },
 }
 
-module.exports = new Split();
+module.exports = Split;